Finish the self-contained project setup.
- Add bytecodetranslator submodule (which in turn contains corral).  Add
  all source projects to SVAuth.sln so they can all be built at once.
- Move runtime data from C:\SVX to svx-runtime-data in the working tree.
- Add setup instructions to readme.

Concurrency Explorer is still a manual installation for now.

The part in the readme about the ASP.NET relying party application is
wrong at the moment but will be correct very soon when I convert the
relying party code to SVX2 and switch Main back from the simplified
examples to the web server.  At that point, I will also delete the SVX1
code and the obsolete PoirotRoot setting.