Make Export calls happen regardless of BypassCertification.
The conditional was a remnant of when I overloaded BypassCertification to
indicate whether the concrete subclass had adopted the SVX framework at
all.  Now all subclasses have adopted it, and I want to use
BypassCertification purely to skip the certification to shorten the
development cycle.  If we have non-SVX subclasses again in the future, we
can add a separate flag.